BROKEN
Getting some sort of bug with updates where the first update (initial
state) is missed by some consumers. This may be a race condition with
multiple consumers. Seems strange though. Happened all of a sudden after
apparently unrelated changes.

My theory is that this is causing state to be missed, because prev is
probably null.

Yeah, membranes are actually _not getting_ the first update because it's
somehow getting missed. Not sure if this is all membranes.